requirements.txt
cachetools==4.0.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
httplib2==0.17.0
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
robotframework==3.1.2
robotframework-seleniumlibrary==4.3.0
rsa==4.0
selenium==3.141.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
